added new workshop component
